+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.1] - 2025-08-31
+
+### Changed
+ - Switched from front-end API url to *official* REST API url for autocompletion.
+ - Updated documentation for `rule34Py.autocomplete`
+ - Updated unit tests
+ - Changed `API_URLS.AUTOCOMPLETE` to semi official endpoint
 
 ## [4.0.0] - 2025-08-31